open_nodes: introduce edbg base class and use it with samr21 and arduino-zero
