+++ gateway_code/open_nodes/node_arduino_zero.py
@@ -20,117 +20,16 @@
 # knowledge of the CeCILL license and that you accept its terms.
 
 """ Open Node Arduino Zero experiment implementation """
-import logging
 
 from gateway_code.config import static_path
-from gateway_code import common
-from gateway_code.common import logger_call
-
-from gateway_code.nodes import OpenNodeBase
-from gateway_code.utils.openocd import OpenOCD
-from gateway_code.utils.edbg import Edbg
-from gateway_code.utils.serial_redirection import SerialRedirection
-
-LOGGER = logging.getLogger('gateway_code')
+from gateway_code.open_nodes.common.node_edbg import NodeEdbgBase
 
 
-class NodeArduinoZero(OpenNodeBase):
+class NodeArduinoZero(NodeEdbgBase):
     """ Open node Arduino Zero implementation """
 
     TYPE = 'arduino_zero'
-    ELF_TARGET = ('ELFCLASS32', 'EM_ARM')
     TTY = '/dev/iotlab/ttyON_ARDUINO_ZERO'
-    BAUDRATE = 115200
     OPENOCD_CFG_FILE = static_path('iot-lab-arduino-zero.cfg')
     FW_IDLE = static_path('arduino_zero_idle.elf')
     FW_AUTOTEST = static_path('arduino_zero_autotest.elf')
-
-    AUTOTEST_AVAILABLE = [
-        'echo', 'get_time',  # mandatory
-        'leds_on', 'leds_off'
-    ]
-
-    ALIM = '5V'
-
-    def __init__(self):
-        self.serial_redirection = SerialRedirection(self.TTY, self.BAUDRATE)
-        self.openocd = OpenOCD.from_node(self)
-        self.edbg = Edbg()
-        self._current_fw = None
-        self._in_debug = False
-
-    @logger_call("Node Arduino Zero : Setup of arduino zero node")
-    def setup(self, firmware_path):
-        """ Flash open node, create serial redirection """
-        self._in_debug = False
-
-        ret_val = 0
-
-        common.wait_no_tty(self.TTY)
-        ret_val += common.wait_tty(self.TTY, LOGGER)
-        ret_val += self.flash(firmware_path)
-        ret_val += self.serial_redirection.start()
-        return ret_val
-
-    @logger_call("Node Arduino Zero : teardown of arduino zero node")
-    def teardown(self):
-        """ Stop serial redirection and flash idle firmware """
-        self._in_debug = False
-
-        ret_val = 0
-        # ON may have been stopped at the end of the experiment.
-        # And then restarted again in cn teardown.
-        # This leads to problem where the TTY disappears and reappears during
-        # the first 2 seconds. So let some time if it wants to disappear first.
-        common.wait_no_tty(self.TTY)
-        ret_val += common.wait_tty(self.TTY, LOGGER)
-        # cleanup debugger before flashing
-        ret_val += self.debug_stop()
-        ret_val += self.serial_redirection.stop()
-        ret_val += self.flash(None)
-        return ret_val
-
-    @logger_call("Node Arduino Zero : flash of arduino zero node")
-    def flash(self, firmware_path=None):
-        """ Flash the given firmware on Arduino Zero node
-
-        :param firmware_path: Path to the firmware to be flashed on `node`.
-                              If None, flash 'idle' firmware.
-        """
-        firmware_path = firmware_path or self.FW_IDLE
-        LOGGER.info('Flash firmware on Arduino Zero: %s', firmware_path)
-        self._current_fw = firmware_path
-
-        if self._in_debug:
-            return self.openocd.flash(firmware_path)
-
-        return self.edbg.flash(firmware_path)
-
-    @logger_call("Node Arduino Zero : reset of arduino zero node")
-    def reset(self):
-        """ Reset the Arduino Zero node using jtag """
-        LOGGER.info('Reset Arduino Zero node')
-        return self.openocd.reset()
-
-    def debug_start(self):
-        """ Start Arduino Zero node debugger """
-        LOGGER.info('Arduino Zero Node debugger start')
-        self._in_debug = True
-        if self._current_fw is not None:
-            # Reflash current firmware using openocd to avoid misbehavior of
-            # previously flashed firmware (with edbg)
-            self.flash(firmware_path=self._current_fw)
-        return self.openocd.debug_start()
-
-    def debug_stop(self):
-        """ Stop Arduino Zero node debugger """
-        LOGGER.info('Arduino Zero Node debugger stop')
-        self._in_debug = False
-        return self.openocd.debug_stop()
-
-    @staticmethod
-    def status():
-        """ Check Arduino Zero node status """
-        # Status is called when open node is not powered
-        # So can't check for FTDI
-        return 0
